+++ index.js
@@ -69,6 +69,7 @@
       f: bar      Links to Forth.prototype.bar (defaults to same as 'n'
       token: true Create a token word that can be used as first cell of a definition - order must match constants below
       replaced: true  This word will be replaced by a Forth definition, so check for its being compiled into other definitions.
+      jsNeeds: true The execution address of this word is needed by the JS o will be stored in jstoxt[n]
  */
 // it is used to build the dictionary in each instance.
 const jsFunctionAttributes = [
@@ -163,7 +164,7 @@ USER(undefined, 0);    // Always 0 to indicate no more valid Vocabularies
 const CURRENToffset = USER('CURRENT', undefined); // Point to the vocabulary to be extended. Default to FORTH.
 // this is set to result of currentFetch
 USER(undefined, undefined); // Vocabulary link uses one cell after CURRENT (not clear how this is used)
-// Note there is a (hidden) assumption that CP and NP are consecutive - used to check free space
+// Not e there is a (hidden) assumption that CP and NP are consecutive - used to check free space
 const CPoffset = USER('CP', undefined);  // eForth initializes to CTOP but we have to use this during compilation
 const NPoffset = USER('NP', undefined);  // normally set on Zen pg106 but we are using it live. Its the bottom of what compiled in name space.
 const LASToffset = USER('LAST', undefined); // normally set on Zen pg106 but using live
@@ -179,42 +180,51 @@ const forthInForth = `
 : IMMEDIATE =IMED setHeaderBits ;
 
 : ( 41 PARSE 2DROP ; IMMEDIATE ( from Zen pg74, note dont have LITERAL so cant do [ CHAR ( ] LITERAL instead of 41 )
-  1 2 ( 3 ) 1 2 2 TEST
-( Now that ( is defined we can use comments below here )
+: \\ ( -- ; Ignore following text till the end of line.)
+   #TIB @ >IN ! ( store the length of TIB in >IN )
+   ; IMMEDIATE ( in effect ignore the rest of a line )
+( Now these are defined we can use both kinds of comments below here )
+
+1 2 ( 3 ) 1 2 2 TEST
+
+( Uncomment first for production, 2nd for testing )
+: \\T #TIB @ >IN ! ;
+\\ : \\T ; IMMEDIATE
+
 
 ( === Test as many of the words defined in code as possible)
 ( EXIT & EXECUTE tested with ' )
 ( doLIT implicitly tested by all literals )
 ( next, ?branch, branch implicitly tested by control structures )
 ( CELLL 2 1 TEST ( Assuming small cell, otherwise it should be 2 )
-123 HLD ! HLD @ 123 1 TEST ( Also tests user variables )
-222 HLD C! HLD C@ 222 1 TEST
+\\T 123 HLD ! HLD @ 123 1 TEST ( Also tests user variables )
+\\T 222 HLD C! HLD C@ 222 1 TEST
 ( R> >R R@ SP@ SP! tested after arithmetic operators )
-30 20 DROP 30 1 TEST
-30 20 DUP 30 20 20 3 TEST
-10 20 SWAP 20 10 2 TEST
-10 20 OVER 10 20 10 3 TEST
-123 0< -123 0< 0 -1 2 TEST
-16 BASE !
-5050 6060 AND 5050 6060 OR 5050 6060 XOR 4040 7070 3030 3 TEST
-0A BASE !
+\\T 30 20 DROP 30 1 TEST
+\\T 30 20 DUP 30 20 20 3 TEST
+\\T 10 20 SWAP 20 10 2 TEST
+\\T 10 20 OVER 10 20 10 3 TEST
+\\T 123 0< -123 0< 0 -1 2 TEST
+\\T 16 BASE !
+\\T 5050 6060 AND 5050 6060 OR 5050 6060 XOR 4040 7070 3030 3 TEST
+\\T 0A BASE !
 
-12 34 UM+ 46 0 2 TEST
+\\T 12 34 UM+ 46 0 2 TEST
 ( Next test is contrived to work with any CELLL size )
--1 -1 UM+ -2 1 2 TEST
+\\T -1 -1 UM+ -2 1 2 TEST
 ( IMMEDIATE implicitly tested )
-32 PARSE ABC SWAP DROP 3 1 TEST
-!IO 62 TX! 0 TEST ( Should output > )
-: FOO 10 EXIT 20 ; FOO 10 1 TEST
-' FOO EXECUTE 30 10 30 2 TEST
+\\T 32 PARSE ABC SWAP DROP 3 1 TEST
+\\T !IO 62 TX! 0 TEST ( Should output > )
+\\T : FOO 10 EXIT 20 ; FOO 10 1 TEST
+\\T ' FOO EXECUTE 30 10 30 2 TEST
 
 
 ( === First section is words moved earlier than they appear in eForth as they are needed for other words )
 ( they are mostly still in the same order as in eForth )
 
-BL 32 1 TEST
+\\T BL 32 1 TEST
 : NIP SWAP DROP ;
-1 2 NIP 2 1 TEST
+\\T 1 2 NIP 2 1 TEST
 
 ( Note failure on this line, can be for three reasons )
 ( a: its the first execution of an immediate word inside a compilation )
@@ -282,9 +292,9 @@ BL 32 1 TEST
 : 2DUP OVER OVER ; ( w1 w2 -- w1 w2 w1 w2;)
 ( 2DROP & NIP moved earlier )
 
-1 ?DUP 0 ?DUP 1 1 0 3 TEST
-1 2 3 ROT 2 3 1 3 TEST
-1 2 2DUP 1 2 1 2 4 TEST
+\\T 1 ?DUP 0 ?DUP 1 1 0 3 TEST
+\\T 1 2 3 ROT 2 3 1 3 TEST
+\\T 1 2 2DUP 1 2 1 2 4 TEST
 
 ( === More Arithmetic operators Zen pg50 )
 : + UM+ DROP ; ( w1 w2 -- w1+w2)
@@ -297,14 +307,14 @@ BL 32 1 TEST
 : ABS DUP 0< IF NEGATE THEN ; ( n -- n; Absolute value of w)
 : 0= IF 0 ELSE -1 THEN ;
 
-1 2 + 3 1 TEST
--1 1 -1 3 D+ -2 5 2 TEST ( Test contrived to be CELLL independent )
-257 INVERT -258 1 TEST ( 257 is 0x101 65278 is 0xFEFE )
-53 NEGATE 53 + 0 1 TEST
-456 123 - 456 -123 - 333 579 2 TEST
--456 ABS 456 ABS - 0 1 TEST
-111 >R R@ RP@ R> SWAP RP@ SWAP - 111 111 CELLL 3 TEST
-1 2 SP@ CELL+ SP! 1 1 TEST
+\\T 1 2 + 3 1 TEST
+\\T -1 1 -1 3 D+ -2 5 2 TEST ( Test contrived to be CELLL independent )
+\\T 257 INVERT -258 1 TEST ( 257 is 0x101 65278 is 0xFEFE )
+\\T 53 NEGATE 53 + 0 1 TEST
+\\T 456 123 - 456 -123 - 333 579 2 TEST
+\\T -456 ABS 456 ABS - 0 1 TEST
+\\T 111 >R R@ RP@ R> SWAP RP@ SWAP - 111 111 CELLL 3 TEST
+\\T 1 2 SP@ CELL+ SP! 1 1 TEST
 
 ( === More comparison Zen pg51-52 )
 : = XOR IF 0 EXIT THEN -1 ; ( w w -- t)
@@ -314,12 +324,12 @@ BL 32 1 TEST
 : MIN 2DUP SWAP < IF SWAP THEN DROP ;
 : WITHIN OVER - >R - R> U< ;
 
-123 123 = 123 124 = -1 0 2 TEST
-123 100 U< 100 123 U< 123 -100 U< -100 123 U< 0 -1 -1 0 4 TEST
-123 100 < 100 123 < 123 -100 < -100 123 < 0 -1 0 -1 4 TEST
-100 200 MAX 300 100 MAX 200 300 2 TEST
-100 200 MIN 300 100 MIN 100 100 2 TEST
-200 100 300 WITHIN 300 100 200 WITHIN 100 -100 200 WITHIN -1 0 -1 3 TEST
+\\T 123 123 = 123 124 = -1 0 2 TEST
+\\T 123 100 U< 100 123 U< 123 -100 U< -100 123 U< 0 -1 -1 0 4 TEST
+\\T 123 100 < 100 123 < 123 -100 < -100 123 < 0 -1 0 -1 4 TEST
+\\T 100 200 MAX 300 100 MAX 200 300 2 TEST
+\\T 100 200 MIN 300 100 MIN 100 100 2 TEST
+\\T 200 100 300 WITHIN 300 100 200 WITHIN 100 -100 200 WITHIN -1 0 -1 3 TEST
 
 ( === More Math Words Zen pg53-55 UM/MOD M/MOD /MOD MOD / UM+ * M* )
 : UM/MOD ( udl udh u -- ur uq ) ( needs FOR-NEXT from 91)
@@ -397,15 +407,15 @@ BL 32 1 TEST
   ABS SWAP ABS UM*  ( multiply absolutes)
   R> IF DNEGATE THEN ;  ( negate if signs are different)
 
--1 1 2 UM/MOD 1 -1 2 TEST
--1 1 2 M/MOD 1 -1 2 TEST
-9 0 4 UM/MOD 1 2 2 TEST
-9 4 /MOD 1 2 2 TEST
-9 4 MOD 1 1 TEST
-9 4 / 2 1 TEST
-2 -1 UM* -2 1 2 TEST
-3 4 * 12 1 TEST
--2 -3 M* 6 0 2 TEST
+\\T -1 1 2 UM/MOD 1 -1 2 TEST
+\\T -1 1 2 M/MOD 1 -1 2 TEST
+\\T 9 0 4 UM/MOD 1 2 2 TEST
+\\T 9 4 /MOD 1 2 2 TEST
+\\T 9 4 MOD 1 1 TEST
+\\T 9 4 / 2 1 TEST
+\\T 2 -1 UM* -2 1 2 TEST
+\\T 3 4 * 12 1 TEST
+\\T -2 -3 M* 6 0 2 TEST
 
 ( === Scaling Words Zen pg56 */MOD */ )
 
@@ -418,8 +428,8 @@ BL 32 1 TEST
   ( Multiple by n1 by n2, then divide by n3. Return quotient only)
   */MOD NIP ; ( n1*n2/n3 and discard remainder)
 
-5 7 2 */MOD 1 17 2 TEST
-5 7 2 */ 17 1 TEST
+\\T 5 7 2 */MOD 1 17 2 TEST
+\\T 5 7 2 */ 17 1 TEST
 
 ( === Memory Alignment words Zen pg57 CELL+ CELL- CELLS ALIGNED )
 ( CELL+ & ALIGNED moved earlier ERRATA Zen & v5 use 2 which is wrong unless CELL is '2' )
@@ -432,8 +442,8 @@ BL 32 1 TEST
   ( Multiply n by cell size in bytes)
   CELLL * ;
 
-123 CELL- CELLL + 123 1 TEST
-123 CELLS CELLL / 123 1 TEST
+\\T 123 CELL- CELLL + 123 1 TEST
+\\T 123 CELLS CELLL / 123 1 TEST
 
 ( === Special Characters Zen pg58 BL >CHAR )
 
@@ -444,7 +454,7 @@ BL 32 1 TEST
   127 BL WITHIN ( if its a control character)
   IF DROP 95 THEN ; ( replace with an underscore)
 
-41 >CHAR 23 >CHAR BL >CHAR 41 95 32 3 TEST
+\\T 41 >CHAR 23 >CHAR BL >CHAR 41 95 32 3 TEST
 
 ( === Managing Data Stack Zen pg59 DEPTH PICK )
 
@@ -460,8 +470,8 @@ BL 32 1 TEST
   1 + CELLS     ( bytes below tos)
   SP@ + @ ;     ( fetch directly from stack)
 
-1 2 3 DEPTH 1 2 3 3 4 TEST
-11 22 33 1 PICK 11 22 33 22 4 TEST
+\\T 1 2 3 DEPTH 1 2 3 3 4 TEST
+\\T 11 22 33 1 PICK 11 22 33 22 4 TEST
 
 ( === Memory Access Zen pg60 +! 2! 2@ HERE PAD TIB @EXECUTE )
 ( Note 2! 2@ presume big-endian which is a Forth assumption (high word at low address)
@@ -491,10 +501,10 @@ BL 32 1 TEST
   ELSE break
   THEN ;
 
-HLD @ 2 HLD +! HLD @ SWAP - 2 1 TEST
-1 2 3 SP@ 4 5 ROT 2! 1 4 5 3 TEST
-1 2 3 SP@ 2@ 1 2 3 2 3 5 TEST
-TIB >R BL PARSE XXX SWAP R> - 3 16 2 TEST
+\\T HLD @ 2 HLD +! HLD @ SWAP - 2 1 TEST
+\\T 1 2 3 SP@ 4 5 ROT 2! 1 4 5 3 TEST
+\\T 1 2 3 SP@ 2@ 1 2 3 2 3 5 TEST
+\\T TIB >R BL PARSE XXX SWAP R> - 3 19 2 .S TEST
 
 ( === Memory Array and String Zen pg61-62: COUNT CMOVE FILL -TRAILING PACK$ )
 
@@ -544,7 +554,7 @@ TIB >R BL PARSE XXX SWAP R> - 3 16 2 TEST
   1 + SWAP CMOVE      ( ; store characters in cells - 0 padded to end of cell)
   R> ;        ( leave only word buffer address )
 
-  NP @ CELL+ CELL+ COUNT NIP 5 1 TEST
+\\T NP @ CELL+ CELL+ COUNT NIP 5 1 TEST
   ( TODO-11-CELLL - need these tests)
   ( TODO rework this  test('NP @ 4 + COUNT PAD SWAP CMOVE', [], {pad: 'PACK$'});
   ( TODO rework this  test('PAD 3 + 5 BL FILL', [], {pad: 'PAC     '});
@@ -562,7 +572,7 @@ TIB >R BL PARSE XXX SWAP R> - 3 16 2 TEST
   0 SWAP UM/MOD   ( divide n by base)
   SWAP DIGIT ;    ( convert remainder to a digit)
 
-123 10 EXTRACT 12 51 2 TEST
+\\T 123 10 EXTRACT 12 51 2 TEST
 
 ( === Number formatting Zen pg65 <# HOLD #S SIGN #> )
 
@@ -611,7 +621,7 @@ TIB >R BL PARSE XXX SWAP R> - 3 16 2 TEST
   FOR AFT DUP C@ EMIT 1 + THEN NEXT DROP ;
 : .$ ( a -- ) COUNT TYPE ; ( from Staapl, not in eForth)
 
-60 EMIT SPACE 2 SPACES 61 EMIT 0 TEST
+\\T 60 EMIT SPACE 2 SPACES 61 EMIT 0 TEST
 ( .$ is tested by ."| and others )
 
 ( === Number output Zen pg66 str HEX DECIMAL .R U.R U. . ? )
@@ -651,7 +661,7 @@ TIB >R BL PARSE XXX SWAP R> - 3 16 2 TEST
 : ? ( a -- ; Display the contents in a memory cell.)
   @ . ; ( very simple but useful command)
 
--123 5 .R 123 5 U.R 123 U. 123 . BASE ? 0 TEST
+\\T -123 5 .R 123 5 U.R 123 U. 123 . BASE ? 0 TEST
 
 ( === Numeric input Zen pg67-68 DIGIT? NUMBER? )
 ( ERRATA Zen NIP is used but not defined )
@@ -709,10 +719,10 @@ TIB >R BL PARSE XXX SWAP R> - 3 16 2 TEST
   2DROP             ( discard garbage )
   R> BASE ! ;       ( restore radix )
 
-50 10 DIGIT? 2 -1 2 TEST
-BL PARSE 1234 PAD PACK$ NUMBER? DROP 1234 1 TEST
+\\T 50 10 DIGIT? 2 -1 2 TEST
+\\T BL PARSE 1234 PAD PACK$ NUMBER? DROP 1234 1 TEST
 ' NUMBER? 'NUMBER !
-123 123 1 TEST
+\\T 123 123 1 TEST
 
 ( TODO Note the EFORTH-ZEN-ERRATA in the docs v. code for NUMBER? means we drop testing the flag will reconsider when see how used )
 ( === Serial I/O Zen pg69 ?KEY KEY EMIT NUF? )
@@ -827,17 +837,13 @@ BL PARSE 1234 PAD PACK$ NUMBER? DROP 1234 1 TEST
 
 : CHAR BL PARSE DROP C@ ; ( Parse a word and return its first character )
 : CTRL CHAR 31 AND ; ( Parse a word, return first character as a control )
-CHAR ) 41 1 TEST
-CTRL H 8 1 TEST
+\\T CHAR ) 41 1 TEST
+\\T CTRL H 8 1 TEST
 
 : .(  ( -- ) ( Output following string up to next )
   [ CHAR ) ] LITERAL PARSE
    ( parse the string until next )
   TYPE ; IMMEDIATE         ( type the string to terminal )
-
-: \\ ( -- ; Ignore following text till the end of line.)
-   #TIB @ >IN ! ( store the length of TIB in >IN )
-   ; IMMEDIATE ( in effect ignore the rest of a line )
 
 ( Note there is this.TOKEN which does same thing )
 : TOKEN ( -- a ; <string> ; Parse a word from input stream and copy it to name dictionary.)
@@ -850,11 +856,11 @@ CTRL H 8 1 TEST
   PARSE         ( parse out a string delimited by c )
   HERE PACK$ ;  ( copy the string into the word buffer )
 
-1 2 ( 3 ) 4 \\ commenting
-1 2 4 3 TEST
-: foo TOKEN C@ ; foo xxx 3 1 TEST
-BL WORD yyyy DUP C@ SWAP CP @ - 4 0 2 TEST
-: foo .( output at compile time) ; 0 TEST
+\\T 1 2 ( 3 ) 4 \\ commenting
+\\T 1 2 4 3 TEST
+\\T : foo TOKEN C@ ; foo xxx 3 1 TEST
+\\T BL WORD yyyy DUP C@ SWAP CP @ - 4 0 2 TEST
+\\T : foo .( output at compile time) ; 0 TEST
 
 ( === Dictionary Search Zen pg75-77 NAME> SAME? find NAME? )
 
@@ -942,9 +948,9 @@ BL WORD yyyy DUP C@ SWAP CP @ - 4 0 2 TEST
   0 ;           ( exit with a false flag )
 
 ( NAME> implicitly tested by find )
-BL WORD xxx DUP C@ 1 + PAD SWAP CMOVE PAD BL WORD xxx 4 CELLL / SAME? >R 2DROP R> 0 1 TEST
-BL WORD xxx DUP C@ 1 + PAD SWAP CMOVE PAD BL WORD xzx 4 CELLL / SAME? >R 2DROP R> 0= 0 1 TEST
-FORTH 0 TEST
+\\T BL WORD xxx DUP C@ 1 + PAD SWAP CMOVE PAD BL WORD xxx 4 SAME? >R 2DROP R> 0 1 TEST
+\\T BL WORD xxx DUP C@ 1 + PAD SWAP CMOVE PAD BL WORD xzx 4 SAME? >R 2DROP R> 0= 0 1 TEST
+\\T FORTH 0 TEST
 ( TODO convert test now dont have testFind ; BL WORD TOKEN CONTEXT @ FORTHfind', testFind.map(k => k ; (Compare with results from old version of find )
 ( TODO convert this test - now dont have testFind - BL WORD TOKEN NAME?', testFind ; // Name searches all vocabs )
 
@@ -1081,11 +1087,11 @@ CREATE NULL$ 0 , ( EFORTH-ZEN-ERRATA inserts a string "coyote" after this, no id
   $," ; IMMEDIATE ( compile print string )
 
 ( Test is tricky - the "3" in bar is thrown away during hte "THROW" while 5 is argument to THROW )
-: bar 3 5 THROW 4 ; : foo 1 [ ' bar ] LITERAL CATCH 2 ; foo 1 5 2 3 TEST
-: bar 3 ; : foo 1 [ ' bar ] LITERAL CATCH 2 ; foo 1 3 0 2 4 TEST
-: foo ." hello" ; foo 0 TEST
+\\T : bar 3 5 THROW 4 ; : foo 1 [ ' bar ] LITERAL CATCH 2 ; foo 1 5 2 3 TEST
+\\T : bar 3 ; : foo 1 [ ' bar ] LITERAL CATCH 2 ; foo 1 3 0 2 4 TEST
+\\T : foo ." hello" ; foo 0 TEST
 ( Note that abort restores the stack, so shouldn't have consumed something else will have random noise on stack )
-: bar ?DUP ABORT" test" 3 ; : foo [ ' bar ] LITERAL CATCH ; 1 foo C@ 0 foo 1 4 3 0 4 TEST
+\\T : bar ?DUP ABORT" test" 3 ; : foo [ ' bar ] LITERAL CATCH ; 1 foo C@ 0 foo 1 4 3 0 4 TEST
 ( $," and abort" are implicitly tested by ABORT")
 
 ( === Text Interpreter loop Zen pg83-84 $INTERPRET [ .OK ?STACK EVAL )
@@ -1128,9 +1134,9 @@ CREATE NULL$ 0 , ( EFORTH-ZEN-ERRATA inserts a string "coyote" after this, no id
 
 ( === TODO-TEST TODO-IO test EVAL, not that .OK wont work here since not yet using $INTERPRET )
 
-BL PARSE 123 PAD PACK$ $INTERPRET 123 1 TEST
-123 BL PARSE DUP PAD PACK$ $INTERPRET  123 123 2 TEST
-: foo 1 2DROP ?STACK ; : bar [ ' foo ] LITERAL CATCH ; bar C@ 9 1 TEST
+\\T BL PARSE 123 PAD PACK$ $INTERPRET 123 1 TEST
+\\T 123 BL PARSE DUP PAD PACK$ $INTERPRET  123 123 2 TEST
+\\T : foo 1 2DROP ?STACK ; : bar [ ' foo ] LITERAL CATCH ; bar C@ 9 1 TEST
 
 ( This switches to use new interpreter, its still using old js $COMPILE )
 
@@ -1140,7 +1146,7 @@ BL PARSE 123 PAD PACK$ $INTERPRET 123 1 TEST
   'EVAL !                   ( store $INTERPRET in 'EVAL )
   ; IMMEDIATE               ( must be done even while compiling )
 
-[ 1 2 3 ROT 2 3 1 3 TEST
+\\T [ 1 2 3 ROT 2 3 1 3 TEST
 
 ( TODO-ZEN-V5-STAAPL - COMPARE ABOVE HERE)
 
@@ -1189,6 +1195,17 @@ vCREATE I/O  ' ?RX v, ' TX! v, ( Array to store default I/O vectors. )
   QUERY ( get a line of commands from )
   EVAL ; ( Evaluate it)
 
+: quitError ( f -- )
+  ( Handle a possible error returned by EVAL - common to QUIT and quit1 )
+      NULL$ OVER XOR      ( is error address=NULL$ ? )
+    ( V5, ZEN and Staapl differ, prefer Staapl I think)
+    IF                  ( its not NULL$ )
+      CR TIB #TIB @ TYPE ( Display line in TIB )
+      CR >IN @ [ CHAR ^ ] LITERAL CHARS ( ^ under offending word )
+      CR .$ ."  ? "     ( followed by error message and "?" )
+    THEN
+    PRESET ;             ( reset the data stack )
+
 : QUIT ( -- )
  ( Reset return stack pointer and start text interpreter. )
   RP0 @ RP!           ( initialize the return stack )
@@ -1201,16 +1218,19 @@ vCREATE I/O  ' ?RX v, ' TX! v, ( Array to store default I/O vectors. )
     UNTIL ( a)          ( exit if an error occurred )
     ( 'PROMPT @ SWAP )  ( EFORTH-ZEN and EFORTH-V5 save current prompt address, Staapl doesnt)
     CONSOLE             ( Initialize for terminal interaction)
-    NULL$ OVER XOR      ( is error address=NULL$ ? )
-    ( V5, ZEN and Staapl differ, prefer Staapl I think)
-    IF                  ( its not NULL$ )
-      CR TIB #TIB @ TYPE ( Display line in TIB )
-      CR >IN @ [ CHAR ^ ] LITERAL CHARS ( ^ under offending word )
-      CR .$ ."  ? "     ( followed by error message and "?" )
-    THEN
-    PRESET              ( reset the data stack )
+    quitError           ( Report error and reset data stack) 
     ( V5 and ZEN also send "ERR" to file handler if prompt is not OK which is a little off )
   AGAIN ;               ( go back get another command line )
+
+: quit1 ( -- )
+  ( Evaluate a line that is already in TIB - could be compiling or interpreting, if error then report it)
+  [ ' EVAL ] LITERAL ( evaluate one line)
+  CATCH             ( execute commands with error handler)
+  ?DUP
+  IF ( Its an error )
+    quitError           ( Report error and reset data stack) 
+    [COMPILE] [       ( Exit compiling e.g. if error in a definition )
+  THEN ;
 
 ( TODO-IO test PRESET HAND CONSOLE QUIT )
 
@@ -1231,10 +1251,10 @@ vCREATE I/O  ' ?RX v, ' TX! v, ( Array to store default I/O vectors. )
 ( ERRATA Staapl & Zen do 'CURRENT @ !', v5 does 'NAME> ,' this is fundamentally different usage, ANS matches latter )
 : RECURSE ( -- ) LAST @ NAME> , ; IMMEDIATE
 
-1 ' DUP EXECUTE 1 1 2 TEST
-HERE 2 ALLOT HERE SWAP - 2 1 TEST
-: foo [COMPILE] ( ; foo 2 ) 0 TEST
-: foo ?DUP IF DUP 1 - RECURSE THEN ; 3 foo 3 2 1 3 TEST
+\\T 1 ' DUP EXECUTE 1 1 2 TEST
+\\T HERE 2 ALLOT HERE SWAP - 2 1 TEST
+\\T : foo [COMPILE] ( ; foo 2 ) 0 TEST
+\\T : foo ?DUP IF DUP 1 - RECURSE THEN ; 3 foo 3 2 1 3 TEST
 
 ( === Control Structures Zen pg91-92: FOR BEGIN NEXT UNTIL AGAIN IF AHEAD REPEAT THEN AFT ELSE WHILE )
 ( All moved earlier )
@@ -1247,7 +1267,7 @@ HERE 2 ALLOT HERE SWAP - 2 1 TEST
   COMPILE $"|     ( compile string runtime code)
   $," ; IMMEDIATE ( compile string itself )
 
-: foo $" hello" COUNT NIP ; foo 5 1 TEST
+\\T : foo $" hello" COUNT NIP ; foo 5 1 TEST
 
 ( === Name Dictionary Compiler Zen pg94-96: ?UNIQUE $,n $COMPILE OVERT ; ] call, : IMMEDIATE  (see this.dollarCommaN)
 
@@ -1258,7 +1278,7 @@ HERE 2 ALLOT HERE SWAP - 2 1 TEST
   OVER .$       ( with the offending name )
   THEN DROP ;     ( discard token address )
 
-TOKEN foo DUP ?UNIQUE - 0 1 TEST
+\\T TOKEN foo DUP ?UNIQUE - 0 1 TEST
 
 : $,n ( na -- ) ( See also this.dollarCommaN above)
   ( Build a new dictionary name using the string at na.)
@@ -1273,6 +1293,8 @@ TOKEN foo DUP ?UNIQUE - 0 1 TEST
     ( ptr) ! EXIT             ( save code pointer and exit )
   THEN                        ( here if null input )
   $" name" THROW ;            ( this is an error return )
+
+( TODO-ZEN-V5-STAAPL - COMPARE ABOVE HERE )
 
 : $COMPILE ( a -- ) ( Redefining code word js $COMPILE in Forth)
   ( Compile next word to code dictionary as a token or literal.)
@@ -1289,8 +1311,9 @@ TOKEN foo DUP ?UNIQUE - 0 1 TEST
   IF [COMPILE] LITERAL ( successful. compile a literal number )
     EXIT        ( done )
   THEN          ( not a number either )
-  COUNT TYPE SPACE ." Not found" ( TODO remove line when have better handling of THROW)
   THROW ;       ( generate an error condition )
+
+( TODO-ZEN-V5-STAAPL - COMPARE BELOW HERE )
   
 : OVERT ( -- ) ( Redefining code word in Forth)
   ( TODO-15-EPROM can't write into Code space to update dict - see also code version)
@@ -1316,7 +1339,7 @@ TOKEN foo DUP ?UNIQUE - 0 1 TEST
   tokenDoList ,
   ] ;
 
-: foo 1 ; foo 1 1 TEST
+\\T : foo 1 ; foo 1 1 TEST
 
 
 ( TODO may need JS equivalent - see if used;  : CALL, ( ca -- ; ( DTC 8086 relative call ; [ =CALL ] LITERAL , HERE CELL+ - , ; )
@@ -1330,18 +1353,18 @@ TOKEN foo DUP ?UNIQUE - 0 1 TEST
 ( Create a new word that doesnt allocate any space, but will push address of that space. )
 ( redefines definitions moved up so that they will use new TOKEN etc)
 : create TOKEN $,n OVERT , 0 , ; 
-: CREATE tokenCreate create ; ( Note the extra field for DOES> to patch - 
+: CREATE tokenCreate create ; ( Note the extra field for DOES> to patch ) ( redefines definition moved up so that it will use new TOKEN etc)
 ( vCREATE is like CREATE but if VP is set it makes space in the writable DATA area) 
 : vCREATE ( -- ; <string> ) VP @ ?DUP 0= IF CREATE ELSE tokenVar create , THEN ; ( Compile pointer to data area )
 
-: foo CREATE 123 , DOES> @ ; foo BAR BAR 123 1 TEST
+\\T : foo CREATE 123 , DOES> @ ; foo BAR BAR 123 1 TEST
 
 : VARIABLE ( -- ; <string> ) vCREATE 0 v, ;
 
 : CONSTANT ( u -- ; <string> ) TOKEN $,n OVERT tokenNextVal , , ;
 
-VARIABLE foo 0 TEST
-12 foo ! 0 TEST
+\\T VARIABLE foo 0 TEST
+\\T 12 foo ! 0 TEST
 \\ TEST-15-EPROM test failing: foo @ 12 1 TEST
 
 ( === Utilities Zen pg98 )
@@ -1351,7 +1374,7 @@ VARIABLE foo 0 TEST
 ( ERRATA Staapl - uses COUNT which is technically correct but poor typing )
 
 : _TYPE ( b u -- )
-  ( Unsigned multiply. Return double product.)
+  ( Send a counted string to output )
   FOR           ( repeat u+1 times )
     AFT         (  skip to THEN the first time )
     DUP C@      ( get one character from b )
@@ -1382,8 +1405,8 @@ VARIABLE foo 0 TEST
   THEN
   DROP R> BASE ! ;      ( restore radix )
 
-( TODO-TEST TODO-IO test DUMP (needs NUF? which needs ?RX )
-( LAST @ 48 DUMP 0 TEST )
+( DUMP needs NUF? which needs ?RX, so will fail test if those are not defined yet )
+\\T LAST @ 48 DUMP 0 TEST
 
 ( === Stack Dump Zen pg100 .S )
 
@@ -1397,7 +1420,7 @@ VARIABLE foo 0 TEST
   NEXT          ( repeat until done )
   ."  <sp" ;    ( print stack pointer )
 
-1 2 .S 1 2 2 TEST
+\\T 1 2 .S 1 2 2 TEST
 
 ( === Stack Checking Zen pg101 !CSP ?CSP )
 
@@ -1411,7 +1434,7 @@ VARIABLE foo 0 TEST
 : ?CSP ( -- ; Check stack pointer matches saved stack pointer )
   SP@ CSP @ XOR ABORT" stack depth" ;
 
-.BASE .FREE 1 !CSP ?CSP 1 1 TEST
+\\T .BASE .FREE 1 !CSP ?CSP 1 1 TEST
 
 ( === Dictionary Dump Zen pg102 .ID WORDS )
 
@@ -1436,7 +1459,7 @@ VARIABLE foo 0 TEST
     DROP
   THEN ;          ( end of vocab exit )
 
-( WORDS 0 TEST ; Commented out as expensive )
+\\T WORDS 0 TEST ( Commented out as expensive )
 
 ( === Search Token Name Zen pg103 >NAME )
 : FORTH>NAME ( ca -- na, F ) (
@@ -1451,9 +1474,9 @@ VARIABLE foo 0 TEST
     THEN NIP ?DUP
   UNTIL NIP NIP EXIT  ( found.  return name address )
   THEN DROP 0 ;       ( end of vocabulary, failure )
-BL WORD DUP NAME? SWAP >NAME = -1 1 TEST ( Test FAST JS version )
+\\T BL WORD DUP NAME? SWAP >NAME = -1 1 TEST ( Test FAST JS version )
 ( : >NAME FORTH>NAME ; ) ( Uncomment to use FORTH version of >NAME )
-BL WORD DUP NAME? SWAP FORTH>NAME = -1 1 TEST
+\\T BL WORD DUP NAME? SWAP FORTH>NAME = -1 1 TEST
 
 ( === The simplest Decompiler Zen pg104 SEE )
 : SEE ( -- ; <string> )
@@ -1477,18 +1500,19 @@ BL WORD DUP NAME? SWAP FORTH>NAME = -1 1 TEST
   THEN
   DROP ;
 
-( SEE FORTH>NAME 0 TEST ( Commented out as expensive TODO seems to crash out)
+\\T SEE FORTH>NAME 0 TEST ( Commented out as expensive TODO seems to crash out)
 
 ( ERRATA Zen uses CONSTANT but doesnt define it )
 ( === Signon Message Zen pg105 VER hi )
 
-6 CONSTANT VER ( Return the version number of this implementation.)
+9 CONSTANT VER ( Return the version number of this implementation.)
 
+: version CR ." webFORTH V" VER <# # # 46 HOLD # # 46 HOLD # #> TYPE ( display sign-on text and version )  CR ;
 ( ERRATA v5 'hi' doesnt restore BASE )
+
 : hi ( -- )
   !IO           ( initialize terminal I/O )
-  CR ." webFORTH V" VER <# # # 46 HOLD # # 46 HOLD # #> TYPE ( display sign-on text and version )
-  CR ; COMPILE-ONLY
+  version ; COMPILE-ONLY
 
 ( === Hardware Reset Zen pg106 COLD )
 : EMPTY ( -- ) ( TODO-15-EPROM)
@@ -1559,18 +1583,35 @@ class Mem8 extends Uint8Array {
 class Mem8_16 extends Mem8 {
   pushCell(a, v) { return this.push16(a, v); }
   fetchCell(a) { return (this[a++] << 8) | this[a]; }
-  // TODO this debug is wrong ! Since 8_16 has big-endian data store
-  debug(start, end) { return (end < start) ? 'UNDERFLOW' : new Uint16Array(this.buffer, start, (end - start) >> 1).toString(); }
+  debug(start, end) {
+    if (end < start) return 'UNDERFLOW';
+    const res = [];
+    let i = start;
+    while (i < end) res.push(((this[i++] << 8) | this[i++]) >>0);
+    return res;
+  }
 }
 class Mem8_24 extends Mem8 {
   pushCell(a, v) { return this.push24(a, v); }
   fetchCell(a) { return (((this[a++] << 8) | this[a++]) << 8) | this[a]; }
-  debug(start, end) { return (end < start) ? 'UNDERFLOW' : new Uint8Array(this.buffer, start, end - start).toString(); }
+  debug(start, end) {
+    if (end < start) return 'UNDERFLOW';
+    const res = [];
+    const i = start;
+    while (i < end) res.push((this[i++]<<16 + this[i++]<<8 + this[i++])>>0);
+    return res;
+  }
 }
 class Mem8_32 extends Mem8 {
   pushCell(a, v) { return this.push32(a, v); }
   fetchCell(a) { return  (((((this[a++] << 8) | this[a++]) << 8) | this[a++]) << 8) | this[a]; }
-  debug(start, end) { return (end < start) ? 'UNDERFLOW' : new Uint32Array(this.buffer, start, (end - start) >> 2).toString(); }
+  debug(start, end) {
+    if (end < start) return 'UNDERFLOW';
+    const res = [];
+    let a = start;
+    while (a < end) res.push((((((this[a++] << 8) | this[a++]) << 8) | this[a++]) << 8) | this[a]);
+    return res;
+  }
 }
 
 // noinspection JSBitwiseOperatorUsage,JSBitwiseOperatorUsage
@@ -1800,7 +1841,7 @@ class Forth {
     // Support parameters for TODO-27-MEMORY TODO-28-MULTI
     // === Support for Debugging ============
 
-    this.debugStack = []; // Maintains a position, like a stack trace, don't manipulate directly use functions below
+    this.debugExcecutionStack = []; // Maintains a position, like a stack trace, don't manipulate directly use functions below
     this.debugName = '?'; // Set in threadtoken()
     this.testing = 0x0; // 0x01 display words passed to interpreters; 0x02 each word in tokenthread - typically set by 'testing3'
     this.testingDepth = 1;
@@ -1863,7 +1904,6 @@ class Forth {
   }
   // Extract some memory writing functions
 
-
   // === Build dictionary, mostly from jsFunctionAttributes
   buildDictionary() {
     // Define the first word in the dictionary, I'm using 'FORTH' for this because we need this variable to define everything else.
@@ -1877,8 +1917,8 @@ class Forth {
     this.OVERT(); // Uses the initialization done by this.Ustore(CURRENToffset) above.
 
     // copy constants over
-    ['CELLL', 'CELLbits', 'CELLMASK', 'TIB0'].forEach(k => this.buildConstant(k, this[k]));
-    Object.entries(l).forEach(kv => this.buildConstant(kv[0], kv[1]));
+    ['CELLL', 'CELLbits', 'CELLMASK', 'TIB0'].forEach((k) => this.buildConstant(k, this[k]));
+    Object.entries(l).forEach((kv) => this.buildConstant(kv[0], kv[1]));
 
     this.js2xt = {};
 
@@ -1899,7 +1939,7 @@ class Forth {
       // regular code functions that just need a pointer.
       } else {
         const xt = this.buildCode(n, tok, attribs);
-        //console.assert(xt === this.JSToXT(n));
+        //console.assert(xt === this.JStoXT(n));
         if (attribs.jsNeeds) { this.js2xt[n] = xt; }
       }
     });
@@ -1907,7 +1947,7 @@ class Forth {
     // Build user variables
     // Bracket with a sanity check - also initializes TIB0 from this.TIB0
     console.assert(this.Ufetch(NPoffset) > 0);
-    jsUsers.forEach(nv => this.buildUser(nv[0], nv[1]));
+    jsUsers.forEach((nv) => this.buildUser(nv[0], nv[1]));
     console.assert(this.Ufetch(NPoffset) > 0);
   }
 
@@ -1930,7 +1970,7 @@ class Forth {
   }
   buildUser(name, init) {
     init = typeof init === 'string'
-      ? (typeof this[init] !== 'undefined' ? this[init] : this.JSToXT(init)) // Either field of this, or function
+      ? (typeof this[init] !== 'undefined' ? this[init] : this.JStoXT(init)) // Either field of this, or function
       : typeof init === 'undefined'
         ? this.Ufetch(this._USER) // Already storing during build (e.g. NP)
         : init;
@@ -1961,16 +2001,23 @@ class Forth {
   }
 
   // === Support for Debugging ============
-  debugClear() { this.debugStack = []; }
-  debugPush() { this.debugStack.push(this.debugName); } // in tokenDoList
-  debugPop() { this.debugStack.pop(); } // in EXIT
+  debugClear() { this.debugExcecutionStack = []; }
+  debugPush() { this.debugExcecutionStack.push(this.debugName); } // in tokenDoList
+  debugPop() { this.debugExcecutionStack.pop(); } // in EXIT
 
+  // Return an array of stack entries
+  debugStack() {
+    return this.m.debug(this.SP, this.SPP);
+  }
+  debugReturnStack() {
+    return this.m.debug(this.RP, this.Ufetch(RP0offset));
+  }
   debugThread(xt) {
     if (this.testing & 0x02) {
       this.debugName = this.xt2name(xt); // Expensive so only done when testing
-      if (this.testingDepth > this.debugStack.length) {
+      if (this.testingDepth > this.debugExcecutionStack.length) {
         //TODO-28-MULTITASK RPP(RP0) and SPP will move
-        console.log('R:', this.Ufetch(RP0offset) === this.RP ? '' : this.m.debug(this.RP, this.Ufetch(RP0offset)), this.debugStack, this.xt2name(xt), 'S:', this.SPP === this.SP ? '' : this.m.debug(this.SP, this.SPP),
+        console.log('R:', this.Ufetch(RP0offset) === this.RP ? '' : this.debugReturnStack(), this.debugExcecutionStack, this.xt2name(xt), 'S:', this.SPP === this.SP ? '' : this.debugStack(),
           this.padTestLength ? ('pad: ' + (this.padTestLength > 0 ? this.m.decodeString(this.padPtr(), this.padPtr() + this.padTestLength) : this.m.decodeString(this.padPtr() + this.padTestLength, this.padPtr()))) : '');
       }
     }
@@ -1978,6 +2025,9 @@ class Forth {
   // === Code words to support debugging on the console
   // Put debugNA in a definition to print a counted string on the console
   debugNA() { console.log('NAME=', this.countedToJS(this.SPfetch())); } // Print the NA on console
+  debugTIB() {
+    return this.m.decodeString(this.Ufetch(TIBoffset) + this.Ufetch(INoffset), this.Ufetch(TIBoffset) + this.Ufetch(nTIBoffset));
+  }
   // Put testing3 in a definition to start outputing stack trace on console.
   testing3() {
     this.testing |= 3; }
@@ -2033,7 +2083,7 @@ class Forth {
   Ufetch(userindex) {
     return this.Mfetch(this.UP + userindex * this.CELLL); }
   Ustore(userindex, w) {
-    return this.Mstore(this.UP + userindex * this.CELLL, w); }
+    this.Mstore(this.UP + userindex * this.CELLL, w); }
 
   // === Access to the USER variables before they are defined
   currentFetch() { return this.Ufetch(CURRENToffset); }
@@ -2128,7 +2178,6 @@ class Forth {
 
   ToNAME() { this.SPpush(this.xt2na(this.SPpop())); }  // Fast version of >NAME see Forth definition below
 
-
   // TODO-29-VOCABULARY This just looks up a in the Context vocabulary, it makes no attempt to use multiple vocabularies
   // If required then fixing this to iterate over the context array should not break anything (this is what NAME? does)
   findName() { // a -- xt na | a F
@@ -2145,17 +2194,25 @@ class Forth {
   }
 
   // Search for a string and return either its XT or 0
-  JSToXT(s) {
+  JStoXT(s, save=false) {
+    let xt = this.js2xt[s];
+    if (xt) return xt; // Quick check for cached
     this.JStoCounted(s);
     this.findName(); // xt na | a F
-    return this.SPpop() ? this.SPpop() : undefined;
+    if (this.SPpop()) {
+      xt = this.SPpop();
+      if (save) this.js2xt[s] = xt;
+      return xt;
+    } else {
+      return undefined;
+    }
   }
 
   // === JS Functions to be able to define words ==== in Zen pg30 these are Macros.
 
   // Compile one or more words into the next consecutive code cells.
   DW(...words) {
-    words.forEach(word => this.Ustore(CPoffset, this.Mstore(this.cpFetch(), word)));
+    words.forEach((word) => this.Ustore(CPoffset, this.Mstore(this.cpFetch(), word)));
   }
 
   // a -- a; Check if a definition of the word at 'a' would be unique and display warning (but continue) if it would not be.
@@ -2301,13 +2358,13 @@ class Forth {
         await maybePromise;
       } else if (!waitFrequency--) {
         //setTimeout(resolve, 0) is same as setImmediate(resolve) but latter is not available in browsers
-        await new Promise(resolve => setTimeout(resolve, 0)); //ASYNC: to allow IO to run
+        await new Promise((resolve) => setTimeout(resolve, 0)); //ASYNC: to allow IO to run
         waitFrequency = 100; // How many cycles to allow a thread swap
       }
     }
   }
   async MS() { // ms --; delay for a period of time.
-    await new Promise(resolve => setTimeout(resolve, this.SPpop()));
+    await new Promise((resolve) => setTimeout(resolve, this.SPpop()));
   }
 
   BYE() { // Should exit all the way out
@@ -2368,6 +2425,7 @@ class Forth {
   doLIT() { this.SPpush(this.IPnext()); }
 
   // See DOES> and CREATE, this patches the field after the token compiled by the create to point to the code following the DOES>
+  // : DOES> R> LAST @ 2 CELLS - @ CELLL + ! ; // Untested Forth version, note side effect of the R> of doing an exit.
   DOES() {
     this.Mstore(
       this.Mfetch(this.Ufetch(LASToffset) - 2 * this.CELLL) + this.CELLL, // field after tokenVar compile by CREATE
@@ -2475,8 +2533,7 @@ class Forth {
     }
   }
 
-
-  // === JS interpreter - could be discarded when done or built out
+  // === Forth words written in JS as needed for the interpreter
 
   PARSE() { // returns b (address) and u (length)
     const delimiter = this.SPpop(); // delimiter
@@ -2576,12 +2633,6 @@ class Forth {
     }
   }
 
-  JStoTIB(s) {
-    const TIBoff = this.Ufetch(TIBoffset);
-    console.assert((TIBoff + s.length) < (this.Ufetch(RP0offset) - 10)); // Check for overlong lines
-    this.Ustore(INoffset, 0); // Start at beginning of TIB
-    this.Ustore(nTIBoffset,  this.m.encodeString(TIBoff, s)); // copy string to TIB, and store length in #TIB
-  }
   async EVAL() { // Same signature as Forth EVAL, reads tokens from TIB and interprets
     while (this.Ufetch(INoffset) < this.Ufetch(nTIBoffset)) {
       this.TOKEN(); // a ; pointing to word in Name Buffer (NB)
@@ -2601,6 +2652,14 @@ class Forth {
     }
   }
 
+  // ==== FORTH Interpreter - words that have no Forth equivalent ====
+  JStoTIB(s) {
+    const TIBoff = this.Ufetch(TIBoffset);
+    console.assert((TIBoff + s.length) < (this.Ufetch(RP0offset) - 10)); // Check for overlong lines
+    this.Ustore(INoffset, 0); // Start at beginning of TIB
+    this.Ustore(nTIBoffset,  this.m.encodeString(TIBoff, s)); // copy string to TIB, and store length in #TIB
+  }
+
   // Take a multiline string, and pass line by line to EVAL
   async interpret(inp) {
     const inputs = inp.split('\n');
@@ -2612,10 +2671,13 @@ class Forth {
         console.log(this.m.debug(this.SP, this.SPP), ' >>', inputline);
       }
       this.JStoTIB(inputline);
-      await this.EVAL(inputline);
+      await this.EVAL();
     }
   }
-
+  async interpret1(inp) {
+    this.JStoTIB(inp);
+    await this.run(this.JStoXT('quit1', true));
+  }
   // === A group of words required for the JS interpreter redefined later
 
   // : [  doLIT $INTERPRET 'EVAL ! ; IMMEDIATE
@@ -2661,13 +2723,13 @@ class Forth {
   }
 
   console() {
-    return this.run(this.JSToXT('WARM'));
+    return this.run(this.JStoXT('WARM'));
   }
   // TODO-29-DOES define DOES> for CREATE-DOES> and tokenDoes - this is not part of eForth, THEN defined Vocabulary as CREATE-DOES word
   //tokenDoes = Forth.tokenFunction(payload => { this.RPpush(this.IP); this.IP = (this.Mfetch8(payload++)<<8)+this.Mfetch8(payload++); this.SPpush(payload++); ); // Almost same as tokenDoList
 }
 const ForthNodeOverrides = {
-  TXbangS: s => process.stdout.write(s),
+  TXbangS: (s) => process.stdout.write(s),
   // Setup I/O to the terminal
   bangIO: () => {
     if (process.stdin.isTTY) {